refactor zk elgamal instruction data logic into a separate file
