+++ token/program-2022/src/extension/confidential_transfer/processor.rs
@@ -35,8 +35,8 @@ use {
     spl_elgamal_registry::state::ElGamalRegistry,
     spl_pod::bytemuck::pod_from_bytes,
     spl_token_confidential_transfer_proof_extraction::{
-        transfer::TransferProofContext, transfer_with_fee::TransferWithFeeProofContext,
-        verify_and_extract_context,
+        instruction::verify_and_extract_context, transfer::TransferProofContext,
+        transfer_with_fee::TransferWithFeeProofContext,
     },
 };
